refactor(routing): explainRoute 를 route-explain.ts 로 분리 — 200줄 경계 준수
라우팅 엔진(query-router)에서 표시 계층을 떼어냈다. 어느 규칙이 왜 이겼는지 보여주는
CLI 전용 출력이라 라우팅 판단과 수명이 다르다. query-router 209→186줄.
