+++ src/lib/cli-format.ts
@@ -179,3 +179,4 @@ export function coerceValue(value: string, type: string): unknown {
     default: return value
   }
 }
+
