+++ src/lib/query-router.ts
@@ -184,26 +184,3 @@ function _mstPipeline(q: string, pattern: Pattern, params: Record<string, unknow
   }
 }
 
-/**
- * 쿼리 의도 분석 결과 (디버깅/로깅용)
- */
-export function explainRoute(query: string): string {
-  const result = routeQuery(query)
-  let explanation = `질의: "${query}"\n`
-  explanation += `도구: ${result.tool}\n`
-  explanation += `근거: ${result.reason}${result.matchedPattern ? ` [${result.matchedPattern}]` : ""}\n`
-  explanation += `파라미터: ${JSON.stringify(result.params, null, 2)}\n`
-
-  if (result.dateRange) {
-    explanation += `날짜범위: ${result.dateRange.from} ~ ${result.dateRange.to}\n`
-  }
-
-  if (result.pipeline) {
-    explanation += `파이프라인:\n`
-    for (const step of result.pipeline) {
-      explanation += `  → ${step.tool}(${JSON.stringify(step.params)})\n`
-    }
-  }
-
-  return explanation
-}
