+++ src/lib/scenario-rules.ts
@@ -13,15 +13,8 @@
 
 /** 지원 시나리오 (tools/scenarios/types.ts 의 ScenarioType 원본) */
 export type ScenarioName =
-  | "penalty"
-  | "customs"
-  | "manual"
-  | "delegation"
-  | "impact"
-  | "timeline"
-  | "compliance"
-  | "time_travel"
-  | "action_plan"
+  | "penalty" | "customs" | "manual" | "delegation" | "impact"
+  | "timeline" | "compliance" | "time_travel" | "action_plan"
 
 export interface ScenarioRule {
   scenario: ScenarioName
@@ -165,9 +158,8 @@ const DECLARED_RULES: ScenarioRule[] = [
   {
     scenario: "compliance",
     hostChain: "chain_ordinance_compare",
-    // chain_ordinance_compare 의 유일한 규칙이라 host 안 경쟁이 없다.
-    // 값은 라우터 우선순위로만 쓰이며 ordinance(5)보다 앞서야 한다 —
-    // "서울시 조례가 상위법에 저촉되나"가 단순 조례 검색에 먹히던 문제(#122)
+    // host 안 유일 규칙이라 값은 라우터 우선순위로만 쓰인다. ordinance(5)보다 앞서야
+    // "서울시 조례가 상위법에 저촉되나"가 단순 조례 검색에 먹히지 않는다(#122)
     precedence: 4,
     patterns: [/적합성|상위법\s*위반|위법|초과|저촉|법제\s*심사/],
     // 판정 어휘(위법·초과 등)는 너무 넓어 그대로 쓰면 모든 질의를 조례비교로 끌어간다
